requirements.txt
streamlit
dotenv
psycopg2
sqlalchemy
PyYAML
pandas
colorama
